fix 64bit race; modify hidecardtag

+++ gframe/ocgapi_types.h
@@ -38,7 +38,10 @@ typedef struct OCG_CardData {
 	int32_t level;
 	/////////kdiy///////////
 	uint32_t attribute;
-	uint32_t race;
+    /////////zdiy///////////
+	//uint32_t race;
+    uint64_t race;
+    /////////zdiy///////////
 	int32_t attack;
 	int32_t defense;
 	uint32_t lscale;
